requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pypdf
sentence-transformers
faiss-cpu
numpy
typing-extensions
python-multipart
geopy
